implement file context at startup
