+++ crustacea/src/language_map.py
@@ -1,0 +1,50 @@
+# ToDo: this should be dynamically derived from the current tree-sitter version
+# https://github.com/grantjenks/py-tree-sitter-languages/blob/a6d4f7c903bf647be1bdcfa504df967d13e40427/repos.txt
+language_map = {
+    ".erl": "erlang",
+    ".lua": "lua",
+    ".el": "elisp",
+    ".mk": "make",           
+    ".dockerfile": "dockerfile",
+    "go.mod": "go-mod",      
+    ".sqlite": "sqlite",
+    ".ex": "elixir",
+    ".elm": "elm",
+    ".kt": "kotlin",
+    ".pl": "perl",
+    ".md": "markdown",
+    ".yaml": "yaml",
+    ".m": "objc",
+    ".sql": "sql",
+    ".hcl": "hcl",
+    ".r": "r",
+    ".dot": "dot",
+    ".hack": "hack",
+    ".rst": "rst",
+    ".lisp": "commonlisp",
+    ".sh": "bash",
+    ".c": "c",
+    ".cs": "c-sharp",
+    ".cpp": "cpp",
+    ".css": "css",
+    ".tmpl": "embedded-template",
+    ".go": "go",
+    ".hs": "haskell",
+    ".html": "html",
+    ".java": "java",
+    ".js": "javascript",
+    ".jsdoc": "jsdoc",
+    ".json": "json",
+    ".jl": "julia",
+    ".ml": "ocaml",
+    ".php": "php",
+    ".py": "python",
+    ".ql": "ql",
+    ".regex": "regex",
+    ".rb": "ruby",
+    ".rs": "rust",
+    ".scala": "scala",
+    ".toml": "toml",
+    ".tsq": "tsq",
+    ".ts": "typescript",
+}
